refactor: reorganize new button components into proper folders
